requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
pyee==9.0.4
python-ffmpeg==1.0.15
typing_extensions==4.4.0
